IAL version 50T1

+++ tools/check_commit_ial.sh
@@ -95,9 +95,11 @@ if [ $(hostname | cut -c 1-7) == 'belenos' -o $(hostname | cut -c 1-7) == 'taran
   gmkpack_l[default]=MIMPIIFC1805
   gmkpack_l[49t0]=IMPIIFC2018
   gmkpack_l[49t2]=IMPIIFCI2302DP
+  gmkpack_l[50t1]=IMPIIFCI2302REPRODP
   gmkpack_o[default]=2y
   gmkpack_o[49t0]=x
   gmkpack_o[49t2]=y
+  gmkpack_o[50t1]=y
   defaultMainPackVersion=01
   defaultRef='split_${cycle}'
   ALLTests="${ALLTests},big_3D"
@@ -107,9 +109,11 @@ else
   gmkpack_l[default]=MPIGFORTRAN920DBL
   gmkpack_l[49t0]=OMPIGFORT920DBL
   gmkpack_l[49t2]=OMPIGFORT920DBL
+  gmkpack_l[50t1]=OMPI5GFORT141DP50
   gmkpack_o[default]=xfftw
   gmkpack_o[49t0]=x
   gmkpack_o[49t2]=x
+  gmkpack_o[50t1]=x
   defaultMainPackVersion=01
   defaultRef='split_${cycle}'
 fi
@@ -327,8 +331,13 @@ if [ $is_directory -eq 1 -o $is_commit -eq 1 ]; then
   if [ $is_directory -eq 1 ]; then
     #The git repository is a directory
     fromdir=$commit
-    content_ial_version=$(scp $commit/src/arome/ial_version.json /dev/stdout 2>/dev/null || echo "")
-    cmd_apl_arome="scp $commit/src/arome/ext/apl_arome.F90 /dev/stdout"
+    if [ -d $commit/src ]; then
+      content_ial_version=$(scp $commit/src/arome/ial_version.json /dev/stdout 2>/dev/null || echo "")
+      cmd_apl_arome="scp $commit/src/arome/ext/apl_arome.F90 /dev/stdout"
+    else
+      content_ial_version=$(scp $commit/ial_version.json /dev/stdout 2>/dev/null || echo "")
+      cmd_apl_arome="scp $commit/ext/apl_arome.F90 /dev/stdout"
+    fi
     packBranch=$(echo $commit | sed 's/\//'${separator}'/g' | sed 's/:/'${separator}'/g' | sed 's/\./'${separator}'/g')
   else
     #The git repository is on github
@@ -450,7 +459,7 @@ if [ $packcreation -eq 1 ]; then
 
     export GMKTMP=/dev/shm
 
-    if [ $cycle == '49t2' ]; then
+    if [ $cycle == '49t2' -o $cycle == '50t1' ]; then
       if ! which ial-git2pack > /dev/null 2>&1; then
         echo "ial-git2pack not found, please install it"
         exit 7
@@ -486,20 +495,27 @@ if [ $packcreation -eq 1 ]; then
       done
       rm -rf $tmpbuilddir
 
-      #Workarounds for 49t2 compilation
       cd $HOMEPACK/$name
-      rm -rf src/local/obstat src/local/oopsifs
-      rm -rf hub/local/src/Atlas hub/local/src/OOPS hub/local/src/ecSDK/?ckit
-      if [ $fullcompilation != 0 ]; then
-        gmkfile=$HOMEPACK/$name/.gmkfile/${gmkpack_l}.*
-        for key in REPRO48 WITH_ODB WITHOUT_ECFLOW; do
-          if ! grep "^MACROS_FRT" $gmkfile | grep -e -D$key; then
-            sed -i "/^MACROS_FRT/s/$/ -D$key/" $gmkfile
+      if [ $cycle == '49t2' ]; then
+        #Workarounds for 49t2 compilation
+        rm -rf src/local/obstat src/local/oopsifs
+        rm -rf hub/local/src/Atlas hub/local/src/OOPS hub/local/src/ecSDK/?ckit
+        if [ $fullcompilation != 0 ]; then
+          gmkfile=$HOMEPACK/$name/.gmkfile/${gmkpack_l}.*
+          for key in REPRO48 WITH_ODB WITHOUT_ECFLOW; do
+            if ! grep "^MACROS_FRT" $gmkfile | grep -e -D$key; then
+              sed -i "/^MACROS_FRT/s/$/ -D$key/" $gmkfile
+            fi
+          done
+          if ! grep "^GMK_CMAKE_ectrans" $gmkfile | grep -e -DENABLE_ETRANS=ON; then
+            sed -i "/^GMK_CMAKE_ectrans/s/$/ -DENABLE_ETRANS=ON/" $gmkfile
           fi
-        done
-        if ! grep "^GMK_CMAKE_ectrans" $gmkfile | grep -e -DENABLE_ETRANS=ON; then
-          sed -i "/^GMK_CMAKE_ectrans/s/$/ -DENABLE_ETRANS=ON/" $gmkfile
         fi
+      fi
+      if [ $cycle == '50t1' ]; then
+        #Workarounds for 50t1 compilation
+        rm -rf src/local/oopsifs
+        rm -rf hub/local/src/Atlas hub/local/src/OOPS
       fi
 
       #Prepare PHYEX inclusion
@@ -602,7 +618,7 @@ if [ $packupdate -eq 1 -o $packcreation -eq 1 ]; then
     cd $HOMEPACK/$name/src/local/phyex
 
     if [ $useexpand == 1 ]; then
-      expand_options="--mnhExpand --removeACC"
+      expand_options="--mnhExpand"
     else
       expand_options=""
     fi
@@ -613,13 +629,18 @@ if [ $packupdate -eq 1 -o $packcreation -eq 1 ]; then
       if [[ $commit == arome${separator}* ]]; then
         $prep_code $prepCodeOpts -c $commit PHYEX #This commit is ready for inclusion
       else
-        $prep_code $prepCodeOpts -c $commit $expand_options $subs -m arome PHYEX -- --shumanFUNCtoCALL --removeACC
+        $prep_code $prepCodeOpts -c $commit $subs -m arome PHYEX -- --shumanFUNCtoCALL --removeACC $expand_options
       fi
     else
       echo "Copy $fromdir"
       mkdir PHYEX
-      scp -q -r $fromdir/src PHYEX/
-      $prep_code $prepCodeOpts $expand_options $subs -m arome PHYEX -- --shumanFUNCtoCALL --removeACC
+      if [ -d $fromdir/src ]; then
+        scp -q -r $fromdir/src PHYEX/
+        $prep_code $prepCodeOpts $subs -m arome PHYEX -- --shumanFUNCtoCALL --removeACC $expand_options
+      else
+        scp -q -r $fromdir/* PHYEX/
+        $prep_code $prepCodeOpts PHYEX #Ready for inclusion
+      fi
     fi
     find PHYEX -type f -exec touch {} \; #to be sure a recompilation occurs
     if [ $packupdate -eq 1 ]; then
@@ -720,16 +741,19 @@ if [ $packupdate -eq 1 -o $packcreation -eq 1 ]; then
         else
           reftree='local'
         fi
-        for file in $EXT/*; do
-          extname=`basename $file`
-          loc=`find ../../$reftree/mpa/ -name $extname | sed "s/\/$reftree\//\/local\//g"`
-          nb=`echo $loc | wc -w`
-          if [ $nb -ne 1 ]; then
-            echo "Don't know where $file must be moved, none or several places found!"
-            exit 9
-          fi
-          mvdiff $file $loc
-        done
+        if [ ! -z "$(\ls $EXT)" ]; then
+          #$EXT is not empty
+          for file in $EXT/*; do
+            extname=`basename $file`
+            loc=`find ../../$reftree/mpa/ -name $extname | sed "s/\/$reftree\//\/local\//g"`
+            nb=`echo $loc | wc -w`
+            if [ $nb -ne 1 ]; then
+              echo "Don't know where $file must be moved, none or several places found!"
+              exit 9
+            fi
+            mvdiff $file $loc
+          done
+        fi
       fi
     fi
     rm -rf PHYEX
@@ -867,7 +891,7 @@ if [ $check -eq 1 ]; then
 
       #Files to compare
       if echo $t | grep 'small' > /dev/null; then
-        if [ $cycle == '49t2' ]; then
+        if [ $cycle == '49t2' -o $cycle == '50t1' ]; then
           filestocheck="$filestocheck ${t},conf_tests/$t/ICMSHFPOS+0002:00"
         else
           filestocheck="$filestocheck ${t},conf_tests/$t/ICMSHFPOS+0002:00 ${t},conf_tests/$t/DHFDLFPOS+0002"
